v0.32
Added a new way of managing the blocklist. It's now also following the set-list standard of having a template and a user-list.
The user-list is now created on initializing the setup.
This will make any blocklist created before this point not be used by the program.
I will leave the original model_blocklist.json in for a month to avoid removing any user data. After this I will remove it from the repository.